requirements.txt
beautifulsoup4==4.11.2
lxml==4.9.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
requests==2.28.2
urllib3==1.26.15